+++ src/index.js
@@ -5,8 +5,10 @@ import TextInput from '@/components/input/TextInput.vue';
 import TextInputArea from '@/components/input/TextInputArea.vue';
 import ListBox from '@/components/input/ListBox.vue';
 import ButtonComponent from '@/components/button/ButtonComponent.vue';
-import BasicDialog from './components/dialog/BasicDialog.vue';
-import ConfirmationDialog from './components/dialog/ConfirmationDialog.vue';
+import BasicDialog from '@/components/dialog/BasicDialog.vue';
+import ConfirmationDialog from '@/components/dialog/ConfirmationDialog.vue';
+import RadioButton from '@/components/radio/RadioButton.vue';
+import RadioGroup from '@/components/radio/RadioGroup.vue';
 
 
 const GtsPrint = {
@@ -19,6 +21,8 @@ const GtsPrint = {
     app.component('ListBox', ListBox);
     app.component('BasicDialog', BasicDialog);
     app.component('ConfirmationDialog', ConfirmationDialog)
+    app.component('RadioButton', RadioButton)
+    app.component('RadioGroup', RadioGroup)
   }
 };